Enables font ligatures on the Mac only.
There are issues with using font ligatures with mintty terminal on
windows.  This change enables ligatures on mac only.  They work on the
native Emacs build for windows, but not in the terminal.  The problem
with using Windows native build is that the directory structures don't
play well with WSL and tags.