Fixes issues with using the theme sanityinc-tommorow-bright
- It seemed like a good idea at the time.  Works fine on my mac, but
terminal mode on WSL with mintty was a mess.  Problem is caused by using
italic font settings in the terminal.  The fonts would be garbled because the
terminal doesn't know how to deal with italics.  Removed this mode and restored
my orginal.  Updated coloring to use the tommorow bright colors though so it's not a
complete loss.
- Added a check to general settings for BROWSER to fix new installations